Moved more code from sim to dag_optimizations
